projects/polymarket-bot/requirements.txt
# Polymarket API
py-clob-client>=0.17.0
web3>=6.0.0

# AI/LLM
openai>=1.0.0
anthropic>=0.18.0

# Data & Utils
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
aiohttp>=3.9.0

# Scheduling
apscheduler>=3.10.0

# CLI
typer>=0.9.0
rich>=13.0.0
